Correctness v4: Alembic foundation and fractional marks (C7)
Every score column was Integer while the grading domain has produced floats
since Correctness v2, so partial credit could not be persisted: a 1.5 became a
1, and half a mark came off the student silently. `create_all` could never fix
that -- it creates missing tables and nothing else -- so this phase introduces
the migration mechanism first and then uses it.

Numeric representation: NUMERIC(7, 2), exact decimal rather than binary float,
because a transcript is an exact record and REAL cannot hold 0.1. Scale 2
covers halves, quarters and hundredths; precision 7 is two orders of magnitude
above any real exam total.

The Decimal/float boundary lives in exactly one place. `backend/models/numeric.py`
defines `Marks`, a TypeDecorator that quantises to Decimal on write and returns
float on read, so JSONResponse and all existing float arithmetic keep working
and no router grew a `float(...)` call. `backend/grading/marks.py` holds the
rules themselves and is provider-neutral: no SDK, FastAPI, SQLAlchemy or
backend.models import, asserted by an AST-level test. A human grader, an open
model or an ensemble writes marks through the same path.

Six columns migrated: QuestionResponse.marks_obtained, ExamResult.marks_obtained,
Question.max_marks, Submission.grade, Assignment.points_possible and
Exam.points_possible. The two maxima are included deliberately -- leaving them
Integer would make an exam out of 37.5 unrepresentable while its questions could
hold fractions.

Migration strategy is baseline + stamp, because deployed databases already
exist and were built by create_all. 0001 reproduces that pre-Alembic schema
exactly, Integer marks columns and all, so `stamp 0001` on an existing database
asserts nothing untrue; 0002 alters the six columns. A fresh database runs
`upgrade head` and gets both. A test pins 0001 to sa.Integer() so it cannot
later be "helpfully" updated and turn every stamp into a lie.

The downgrade refuses rather than truncating: it counts non-integral values in
every affected column and raises, naming table, column and row counts. With
whole-number data it proceeds and is loss-free.

create_all is no longer a second authority. Startup now classifies the database
-- alembic_version present, empty, or pre-Alembic -- and respectively does
nothing, creates and stamps, or warns loudly with the adoption commands. The
stamp degrades to a warning if alembic is unavailable, so startup never
hard-requires migration tooling.

Application compatibility: manual marks go through parse_mark_input, so the
strings the UI posts become numbers and a bad value is a 400 naming the field
rather than a driver error at flush time; update_student_response no longer
nulls a mark when the field is absent; the stats histogram derives its bucket
count instead of assuming points_possible is an int; the grade/max-marks
request models became floats; the marks inputs gained step="any".

C6 is unchanged and re-asserted against the new type: 0.0 is a mark, None is a
missing grading result, and grading_incomplete still blocks finalisation.

Verified on SQLite and statically for PostgreSQL: the migration's offline
PostgreSQL output is asserted to be six ALTER COLUMN ... TYPE NUMERIC(7, 2)
statements with no table drop or recreate, and the model columns are asserted
to compile to NUMERIC(7, 2) on the postgresql dialect. SQLite's dynamic typing
means the behavioural tests alone could not have caught C7; the migration has
NOT been run against a real PostgreSQL server.

283 tests (206 + 77).

+++ backend/routers/exams.py
@@ -184,7 +184,7 @@ async def create_exam(
     class_id: int, 
     title: str = Form(...),
     exam_date: Optional[str] = Form(None),
-    points_possible: Optional[int] = Form(100),
+    points_possible: Optional[float] = Form(100),
     db: AsyncSession = Depends(get_db),
     current_user: User = Depends(get_current_user_required)
 ):  
@@ -513,7 +513,7 @@ async def get_exam_questions(
 class UpdatePartLabels(BaseModel):
     questionId: int
     partLabels: List[str]
-    maxMarks: Optional[int]
+    maxMarks: Optional[float]
 
 class UpdatesPayload(BaseModel):
     updates: List[UpdatePartLabels]
